requirements.txt
pytorch
transformers
numpy
flask
regex
pandas
matplotlib
underthesea
pyvi
vncorenlp
seaborn
tensorflow
pickle
argparse